LPS-53627 move to another package separated from service builder stuff
