+++ portal-service/src/com/liferay/portal/util/BaseSelectableEntry.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.model;
+package com.liferay.portal.util;
 
 import com.liferay.portal.kernel.util.HashUtil;
