+++ portal-service/src/com/liferay/portal/util/SelectableEntry.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.model;
+package com.liferay.portal.util;
 
 import java.io.IOException;
